refactor: Rename `ConsentsConfirmation` > `ConsentsConfirmationPage`
