+++ src/client/pages/ConsentsConfirmationPage.tsx
